replaced the asyncWriter to a simple accumelating based buffer;  the actual logstashwriter should handle async behavior and failover.
